requirements.txt
pytube
GitPython
retry